+++ src/Polly.SharedSpecs/Caching/CacheTResultSpecs.cs
@@ -76,11 +76,15 @@ namespace Polly.Specs.Caching
             ISyncCacheProvider stubCacheProvider = new StubCacheProvider();
             CachePolicy<string> cache = Policy.Cache<string>(stubCacheProvider, TimeSpan.MaxValue);
 
-            stubCacheProvider.Get(operationKey).Should().BeNull();
+            (bool cacheHit1, object fromCache1) = stubCacheProvider.TryGet(operationKey);
+            cacheHit1.Should().BeFalse();
+            fromCache1.Should().BeNull();
 
             cache.Execute(ctx => { return valueToReturn; }, new Context(operationKey)).Should().Be(valueToReturn);
 
-            stubCacheProvider.Get(operationKey).Should().Be(valueToReturn);
+            (bool cacheHit2, object fromCache2) = stubCacheProvider.TryGet(operationKey);
+            cacheHit2.Should().BeTrue();
+            fromCache2.Should().Be(valueToReturn);
         }
 
         [Fact]
@@ -93,7 +97,9 @@ namespace Polly.Specs.Caching
             TimeSpan ttl = TimeSpan.FromMinutes(30);
             CachePolicy<string> cache = Policy.Cache<string>(stubCacheProvider, ttl);
 
-            stubCacheProvider.Get(operationKey).Should().BeNull();
+            (bool cacheHit1, object fromCache1) = stubCacheProvider.TryGet(operationKey);
+            cacheHit1.Should().BeFalse();
+            fromCache1.Should().BeNull();
 
             int delegateInvocations = 0;
             Func<Context, string> func = ctx =>
@@ -108,7 +114,10 @@ namespace Polly.Specs.Caching
             // First execution should execute delegate and put result in the cache.
             cache.Execute(func, new Context(operationKey)).Should().Be(valueToReturn);
             delegateInvocations.Should().Be(1);
-            stubCacheProvider.Get(operationKey).Should().Be(valueToReturn);
+
+            (bool cacheHit2, object fromCache2) = stubCacheProvider.TryGet(operationKey);
+            cacheHit2.Should().BeTrue();
+            fromCache2.Should().Be(valueToReturn);
 
             // Second execution (before cache expires) should get it from the cache - no further delegate execution.
             // (Manipulate time so just prior cache expiry).
@@ -133,11 +142,15 @@ namespace Polly.Specs.Caching
             ISyncCacheProvider stubCacheProvider = new StubCacheProvider();
             CachePolicy<string> cache = Policy.Cache<string>(stubCacheProvider, TimeSpan.Zero);
 
-            stubCacheProvider.Get(operationKey).Should().BeNull();
+            (bool cacheHit1, object fromCache1) = stubCacheProvider.TryGet(operationKey);
+            cacheHit1.Should().BeFalse();
+            fromCache1.Should().BeNull();
 
             cache.Execute(ctx => { return valueToReturn; }, new Context(operationKey)).Should().Be(valueToReturn);
 
-            stubCacheProvider.Get(operationKey).Should().Be(null);
+            (bool cacheHit2, object fromCache2) = stubCacheProvider.TryGet(operationKey);
+            cacheHit2.Should().BeFalse();
+            fromCache2.Should().BeNull();
         }
 
         [Fact]
@@ -210,6 +223,97 @@ namespace Polly.Specs.Caching
 
             cache.Execute(func, new Context("person", new { id = "2" }.AsDictionary())).Should().BeSameAs(person2);
             funcExecuted.Should().BeFalse();
+        }
+
+        #endregion
+
+        #region Caching behaviours, default(TResult)
+
+        [Fact]
+        public void Should_execute_delegate_and_put_value_in_cache_if_cache_does_not_hold_value__default_for_reference_type()
+        {
+            ResultClass valueToReturn = default(ResultClass);
+            const string operationKey = "SomeOperationKey";
+
+            ISyncCacheProvider stubCacheProvider = new StubCacheProvider();
+            CachePolicy<ResultClass> cache = Policy.Cache<ResultClass>(stubCacheProvider, TimeSpan.MaxValue);
+
+            (bool cacheHit1, object fromCache1) = stubCacheProvider.TryGet(operationKey);
+            cacheHit1.Should().BeFalse();
+            fromCache1.Should().BeNull();
+
+            cache.Execute(ctx => { return valueToReturn; }, new Context(operationKey)).Should().Be(valueToReturn);
+
+            (bool cacheHit2, object fromCache2) = stubCacheProvider.TryGet(operationKey);
+            cacheHit2.Should().BeTrue();
+            fromCache2.Should().Be(valueToReturn);
+        }
+
+        [Fact]
+        public void Should_return_value_from_cache_and_not_execute_delegate_if_cache_holds_value__default_for_reference_type()
+        {
+            ResultClass valueToReturnFromCache = default(ResultClass);
+            ResultClass valueToReturnFromExecution = new ResultClass(ResultPrimitive.Good);
+            const string operationKey = "SomeOperationKey";
+
+            ISyncCacheProvider stubCacheProvider = new StubCacheProvider();
+            CachePolicy<ResultClass> cache = Policy.Cache<ResultClass>(stubCacheProvider, TimeSpan.MaxValue);
+            stubCacheProvider.Put(operationKey, valueToReturnFromCache, new Ttl(TimeSpan.MaxValue));
+
+            bool delegateExecuted = false;
+
+            cache.Execute(ctx =>
+            {
+                delegateExecuted = true;
+                return valueToReturnFromExecution;
+            }, new Context(operationKey))
+                .Should().Be(valueToReturnFromCache);
+
+            delegateExecuted.Should().BeFalse();
+        }
+
+        [Fact]
+        public void Should_execute_delegate_and_put_value_in_cache_if_cache_does_not_hold_value__default_for_value_type()
+        {
+            ResultPrimitive valueToReturn = default(ResultPrimitive);
+            const string operationKey = "SomeOperationKey";
+
+            ISyncCacheProvider stubCacheProvider = new StubCacheProvider();
+            CachePolicy<ResultPrimitive> cache = Policy.Cache<ResultPrimitive>(stubCacheProvider, TimeSpan.MaxValue);
+
+            (bool cacheHit1, object fromCache1) = stubCacheProvider.TryGet(operationKey);
+            cacheHit1.Should().BeFalse();
+            fromCache1.Should().BeNull();
+
+            cache.Execute(ctx => { return valueToReturn; }, new Context(operationKey)).Should().Be(valueToReturn);
+
+            (bool cacheHit2, object fromCache2) = stubCacheProvider.TryGet(operationKey);
+            cacheHit2.Should().BeTrue();
+            fromCache2.Should().Be(valueToReturn);
+        }
+
+        [Fact]
+        public void Should_return_value_from_cache_and_not_execute_delegate_if_cache_holds_value__default_for_value_type()
+        {
+            ResultPrimitive valueToReturnFromCache = default(ResultPrimitive);
+            ResultPrimitive valueToReturnFromExecution = ResultPrimitive.Good;
+            valueToReturnFromExecution.Should().NotBe(valueToReturnFromCache);
+            const string operationKey = "SomeOperationKey";
+
+            ISyncCacheProvider stubCacheProvider = new StubCacheProvider();
+            CachePolicy<ResultPrimitive> cache = Policy.Cache<ResultPrimitive>(stubCacheProvider, TimeSpan.MaxValue);
+            stubCacheProvider.Put(operationKey, valueToReturnFromCache, new Ttl(TimeSpan.MaxValue));
+
+            bool delegateExecuted = false;
+
+            cache.Execute(ctx =>
+            {
+                delegateExecuted = true;
+                return valueToReturnFromExecution;
+            }, new Context(operationKey))
+                .Should().Be(valueToReturnFromCache);
+
+            delegateExecuted.Should().BeFalse();
         }
 
         #endregion
@@ -372,7 +476,9 @@ namespace Polly.Specs.Caching
             cache.Invoking(policy => policy.Execute(func, new Context(operationKey), tokenSource.Token))
                 .ShouldThrow<OperationCanceledException>();
 
-            stubCacheProvider.Get(operationKey).Should().BeNull();
+            (bool cacheHit, object fromCache) = stubCacheProvider.TryGet(operationKey);
+            cacheHit.Should().BeFalse();
+            fromCache.Should().BeNull();
         }
 
         #endregion