+++ src/Polly.SharedSpecs/Caching/SerializingCacheProviderAsyncSpecs.cs
@@ -60,13 +60,17 @@ namespace Polly.Specs.Caching
             AsyncSerializingCacheProvider<StubSerialized> serializingCacheProvider = new AsyncSerializingCacheProvider<StubSerialized>(stubCacheProvider.AsyncFor<StubSerialized>(), stubSerializer);
             await serializingCacheProvider.PutAsync(key, objectToCache, new Ttl(TimeSpan.FromMinutes(1)), CancellationToken.None, false);
 
-            serializeInvoked.Should().Be(true);
-            (await stubCacheProvider.GetAsync(key, CancellationToken.None, false)).Should().BeOfType<StubSerialized>()
+            serializeInvoked.Should().BeTrue();
+
+            (bool cacheHit, object fromCache) = await stubCacheProvider.TryGetAsync(key, CancellationToken.None, false);
+
+            cacheHit.Should().BeTrue();
+            fromCache.Should().BeOfType<StubSerialized>()
                 .Which.Original.Should().Be(objectToCache);
         }
 
         [Fact]
-        public async Task Single_generic_SerializingCacheProvider_should_not_serialize_on_put_for_defaultTResult()
+        public async Task Single_generic_SerializingCacheProvider_should_serialize_on_put_for_defaultTResult()
         {
             bool serializeInvoked = false;
             StubSerializer<object, StubSerialized> stubSerializer = new StubSerializer<object, StubSerialized>(
@@ -80,8 +84,13 @@ namespace Polly.Specs.Caching
             AsyncSerializingCacheProvider<StubSerialized> serializingCacheProvider = new AsyncSerializingCacheProvider<StubSerialized>(stubCacheProvider.AsyncFor<StubSerialized>(), stubSerializer);
             await serializingCacheProvider.PutAsync(key, objectToCache, new Ttl(TimeSpan.FromMinutes(1)), CancellationToken.None, false);
 
-            serializeInvoked.Should().Be(false);
-            stubCacheProvider.Get(key).Should().BeNull();
+            serializeInvoked.Should().BeTrue();
+
+            (bool cacheHit, object fromCache) = stubCacheProvider.TryGet(key);
+
+            cacheHit.Should().BeTrue();
+            fromCache.Should().BeOfType<StubSerialized>()
+                .Which.Original.Should().Be(objectToCache);
         }
 
         [Fact]
@@ -100,9 +109,10 @@ namespace Polly.Specs.Caching
             await stubCacheProvider.PutAsync(key, new StubSerialized(objectToCache), new Ttl(TimeSpan.FromMinutes(1)), CancellationToken.None, false);
 
             AsyncSerializingCacheProvider<StubSerialized> serializingCacheProvider = new AsyncSerializingCacheProvider<StubSerialized>(stubCacheProvider.AsyncFor<StubSerialized>(), stubSerializer);
-            object fromCache = await serializingCacheProvider.GetAsync(key, CancellationToken.None, false);
+            (bool cacheHit, object fromCache) = await serializingCacheProvider.TryGetAsync(key, CancellationToken.None, false);
 
-            deserializeInvoked.Should().Be(true);
+            cacheHit.Should().BeTrue();
+            deserializeInvoked.Should().BeTrue();
             fromCache.Should().Be(objectToCache);
         }
 
@@ -117,12 +127,13 @@ namespace Polly.Specs.Caching
             var stubCacheProvider = new StubCacheProvider();
             string key = "some key";
 
-            stubCacheProvider.Get(key).Should().BeNull();
+            stubCacheProvider.TryGet(key).Item1.Should().BeFalse();
 
             AsyncSerializingCacheProvider<StubSerialized> serializingCacheProvider = new AsyncSerializingCacheProvider<StubSerialized>(stubCacheProvider.AsyncFor<StubSerialized>(), stubSerializer);
-            object fromCache = await serializingCacheProvider.GetAsync(key, CancellationToken.None, false);
+            (bool cacheHit, object fromCache) = await serializingCacheProvider.TryGetAsync(key, CancellationToken.None, false);
 
-            deserializeInvoked.Should().Be(false);
+            cacheHit.Should().BeFalse();
+            deserializeInvoked.Should().BeFalse();
             fromCache.Should().Be(default(object));
         }
 
@@ -141,13 +152,17 @@ namespace Polly.Specs.Caching
             AsyncSerializingCacheProvider<StubSerialized> serializingCacheProvider = stubCacheProvider.AsyncFor<StubSerialized>().WithSerializer(stubSerializer);
             await serializingCacheProvider.PutAsync(key, objectToCache, new Ttl(TimeSpan.FromMinutes(1)), CancellationToken.None, false);
 
-            serializeInvoked.Should().Be(true);
-            (await stubCacheProvider.GetAsync(key, CancellationToken.None, false)).Should().BeOfType<StubSerialized>()
+            serializeInvoked.Should().BeTrue();
+
+            (bool cacheHit, object fromCache) = await stubCacheProvider.TryGetAsync(key, CancellationToken.None, false);
+
+            cacheHit.Should().BeTrue();
+            fromCache.Should().BeOfType<StubSerialized>()
                 .Which.Original.Should().Be(objectToCache);
         }
 
         [Fact]
-        public async Task Single_generic_SerializingCacheProvider_from_extension_syntax_should_not_serialize_on_put_for_defaultTResult()
+        public async Task Single_generic_SerializingCacheProvider_from_extension_syntax_should_serialize_on_put_for_defaultTResult()
         {
             bool serializeInvoked = false;
             StubSerializer<object, StubSerialized> stubSerializer = new StubSerializer<object, StubSerialized>(
@@ -161,8 +176,13 @@ namespace Polly.Specs.Caching
             AsyncSerializingCacheProvider<StubSerialized> serializingCacheProvider = stubCacheProvider.AsyncFor<StubSerialized>().WithSerializer(stubSerializer);
             await serializingCacheProvider.PutAsync(key, objectToCache, new Ttl(TimeSpan.FromMinutes(1)), CancellationToken.None, false);
 
-            serializeInvoked.Should().Be(false);
-            stubCacheProvider.Get(key).Should().BeNull();
+            serializeInvoked.Should().BeTrue();
+
+            (bool cacheHit, object fromCache) = stubCacheProvider.TryGet(key);
+
+            cacheHit.Should().BeTrue();
+            fromCache.Should().BeOfType<StubSerialized>()
+                .Which.Original.Should().Be(objectToCache);
         }
 
         [Fact]
@@ -180,9 +200,10 @@ namespace Polly.Specs.Caching
             await stubCacheProvider.PutAsync(key, new StubSerialized(objectToCache), new Ttl(TimeSpan.FromMinutes(1)), CancellationToken.None, false);
 
             AsyncSerializingCacheProvider<StubSerialized> serializingCacheProvider = stubCacheProvider.AsyncFor<StubSerialized>().WithSerializer(stubSerializer);
-            object fromCache = await serializingCacheProvider.GetAsync(key, CancellationToken.None, false);
+            (bool cacheHit, object fromCache) = await serializingCacheProvider.TryGetAsync(key, CancellationToken.None, false);
 
-            deserializeInvoked.Should().Be(true);
+            cacheHit.Should().BeTrue();
+            deserializeInvoked.Should().BeTrue();
             fromCache.Should().Be(objectToCache);
         }
 
@@ -197,12 +218,13 @@ namespace Polly.Specs.Caching
             var stubCacheProvider = new StubCacheProvider();
             string key = "some key";
 
-            stubCacheProvider.Get(key).Should().BeNull();
+            stubCacheProvider.TryGet(key).Item1.Should().BeFalse();
 
             AsyncSerializingCacheProvider<StubSerialized> serializingCacheProvider = stubCacheProvider.AsyncFor<StubSerialized>().WithSerializer(stubSerializer);
-            object fromCache = await serializingCacheProvider.GetAsync(key, CancellationToken.None, false);
+            (bool cacheHit, object fromCache) = await serializingCacheProvider.TryGetAsync(key, CancellationToken.None, false);
 
-            deserializeInvoked.Should().Be(false);
+            cacheHit.Should().BeFalse();
+            deserializeInvoked.Should().BeFalse();
             fromCache.Should().Be(default(object));
         }
 
@@ -257,13 +279,17 @@ namespace Polly.Specs.Caching
             AsyncSerializingCacheProvider<ResultPrimitive, StubSerialized<ResultPrimitive>> serializingCacheProvider = new AsyncSerializingCacheProvider<ResultPrimitive, StubSerialized<ResultPrimitive>>(stubCacheProvider.AsyncFor<StubSerialized<ResultPrimitive>>(), stubTResultSerializer);
             await serializingCacheProvider.PutAsync(key, objectToCache, new Ttl(TimeSpan.FromMinutes(1)), CancellationToken.None, false);
 
-            serializeInvoked.Should().Be(true);
-            (await stubCacheProvider.GetAsync(key, CancellationToken.None, false)).Should().BeOfType<StubSerialized<ResultPrimitive>>()
+            serializeInvoked.Should().BeTrue();
+
+            (bool cacheHit, object fromCache) = await stubCacheProvider.TryGetAsync(key, CancellationToken.None, false);
+
+            cacheHit.Should().BeTrue();
+            fromCache.Should().BeOfType<StubSerialized<ResultPrimitive>>()
                 .Which.Original.Should().Be(objectToCache);
         }
 
         [Fact]
-        public async Task Double_generic_SerializingCacheProvider_should_not_serialize_on_put_for_defaultTResult()
+        public async Task Double_generic_SerializingCacheProvider_should_serialize_on_put_for_defaultTResult()
         {
             bool serializeInvoked = false;
             StubSerializer<ResultPrimitive, StubSerialized<ResultPrimitive>> stubTResultSerializer = new StubSerializer<ResultPrimitive, StubSerialized<ResultPrimitive>>(
@@ -277,8 +303,13 @@ namespace Polly.Specs.Caching
             AsyncSerializingCacheProvider<ResultPrimitive, StubSerialized<ResultPrimitive>> serializingCacheProvider = new AsyncSerializingCacheProvider<ResultPrimitive, StubSerialized<ResultPrimitive>>(stubCacheProvider.AsyncFor<StubSerialized<ResultPrimitive>>(), stubTResultSerializer);
             await serializingCacheProvider.PutAsync(key, objectToCache, new Ttl(TimeSpan.FromMinutes(1)), CancellationToken.None, false);
 
-            serializeInvoked.Should().Be(false);
-            stubCacheProvider.Get(key).Should().BeNull();
+            serializeInvoked.Should().BeTrue();
+
+            (bool cacheHit, object fromCache) = stubCacheProvider.TryGet(key);
+
+            cacheHit.Should().BeTrue();
+            fromCache.Should().BeOfType<StubSerialized<ResultPrimitive>>()
+                .Which.Original.Should().Be(objectToCache);
         }
 
         [Fact]
@@ -296,9 +327,10 @@ namespace Polly.Specs.Caching
             AsyncSerializingCacheProvider<ResultPrimitive, StubSerialized<ResultPrimitive>> serializingCacheProvider = new AsyncSerializingCacheProvider<ResultPrimitive, StubSerialized<ResultPrimitive>>(stubCacheProvider.AsyncFor<StubSerialized<ResultPrimitive>>(), stubTResultSerializer);
 
             await stubCacheProvider.PutAsync(key, new StubSerialized<ResultPrimitive>(objectToCache), new Ttl(TimeSpan.FromMinutes(1)), CancellationToken.None, false);
-            object fromCache = await serializingCacheProvider.GetAsync(key, CancellationToken.None, false);
+            (bool cacheHit, object fromCache) = await serializingCacheProvider.TryGetAsync(key, CancellationToken.None, false);
 
-            deserializeInvoked.Should().Be(true);
+            cacheHit.Should().BeTrue();
+            deserializeInvoked.Should().BeTrue();
             fromCache.Should().Be(objectToCache);
         }
 
@@ -313,12 +345,13 @@ namespace Polly.Specs.Caching
             var stubCacheProvider = new StubCacheProvider();
             string key = "some key";
 
-            stubCacheProvider.Get(key).Should().BeNull();
+            stubCacheProvider.TryGet(key).Item1.Should().BeFalse();
 
             AsyncSerializingCacheProvider<ResultPrimitive, StubSerialized<ResultPrimitive>> serializingCacheProvider = new AsyncSerializingCacheProvider<ResultPrimitive, StubSerialized<ResultPrimitive>>(stubCacheProvider.AsyncFor<StubSerialized<ResultPrimitive>>(), stubTResultSerializer);
-            ResultPrimitive fromCache = await serializingCacheProvider.GetAsync(key, CancellationToken.None, false);
+            (bool cacheHit, ResultPrimitive fromCache) = await serializingCacheProvider.TryGetAsync(key, CancellationToken.None, false);
 
-            deserializeInvoked.Should().Be(false);
+            cacheHit.Should().BeFalse();
+            deserializeInvoked.Should().BeFalse();
             fromCache.Should().Be(default(ResultPrimitive));
         }
 
@@ -338,13 +371,16 @@ namespace Polly.Specs.Caching
                 stubCacheProvider.AsyncFor<StubSerialized<ResultPrimitive>>().WithSerializer(stubTResultSerializer);
             await serializingCacheProvider.PutAsync(key, objectToCache, new Ttl(TimeSpan.FromMinutes(1)), CancellationToken.None, false);
 
-            serializeInvoked.Should().Be(true);
-            (await stubCacheProvider.GetAsync(key, CancellationToken.None, false)).Should().BeOfType<StubSerialized<ResultPrimitive>>()
+            serializeInvoked.Should().BeTrue();
+
+            (bool cacheHit, object fromCache) = await stubCacheProvider.TryGetAsync(key, CancellationToken.None, false);
+            cacheHit.Should().BeTrue();
+            fromCache.Should().BeOfType<StubSerialized<ResultPrimitive>>()
                 .Which.Original.Should().Be(objectToCache);
         }
 
         [Fact]
-        public async Task Double_generic_SerializingCacheProvider_from_extension_syntax_should_not_serialize_on_put_for_defaultTResult()
+        public async Task Double_generic_SerializingCacheProvider_from_extension_syntax_should_serialize_on_put_for_defaultTResult()
         {
             bool serializeInvoked = false;
             StubSerializer<ResultPrimitive, StubSerialized<ResultPrimitive>> stubTResultSerializer = new StubSerializer<ResultPrimitive, StubSerialized<ResultPrimitive>>(
@@ -359,8 +395,13 @@ namespace Polly.Specs.Caching
                 stubCacheProvider.AsyncFor<StubSerialized<ResultPrimitive>>().WithSerializer(stubTResultSerializer);
             await serializingCacheProvider.PutAsync(key, objectToCache, new Ttl(TimeSpan.FromMinutes(1)), CancellationToken.None, false);
 
-            serializeInvoked.Should().Be(false);
-            stubCacheProvider.Get(key).Should().BeNull();
+            serializeInvoked.Should().BeTrue();
+
+            (bool cacheHit, object fromCache) = stubCacheProvider.TryGet(key);
+            
+            cacheHit.Should().BeTrue();
+            fromCache.Should().BeOfType<StubSerialized<ResultPrimitive>>()
+                .Which.Original.Should().Be(objectToCache);
         }
 
         [Fact]
@@ -379,9 +420,10 @@ namespace Polly.Specs.Caching
                 stubCacheProvider.AsyncFor<StubSerialized<ResultPrimitive>>().WithSerializer(stubTResultSerializer);
 
             await stubCacheProvider.PutAsync(key, new StubSerialized<ResultPrimitive>(objectToCache), new Ttl(TimeSpan.FromMinutes(1)), CancellationToken.None, false);
-            ResultPrimitive fromCache = await serializingCacheProvider.GetAsync(key, CancellationToken.None, false);
+            (bool cacheHit, ResultPrimitive fromCache) = await serializingCacheProvider.TryGetAsync(key, CancellationToken.None, false);
 
-            deserializeInvoked.Should().Be(true);
+            cacheHit.Should().BeTrue();
+            deserializeInvoked.Should().BeTrue();
             fromCache.Should().Be(objectToCache);
         }
 
@@ -396,13 +438,14 @@ namespace Polly.Specs.Caching
             var stubCacheProvider = new StubCacheProvider();
             string key = "some key";
 
-            stubCacheProvider.Get(key).Should().BeNull();
+            stubCacheProvider.TryGet(key).Item1.Should().BeFalse();
 
             AsyncSerializingCacheProvider<ResultPrimitive, StubSerialized<ResultPrimitive>> serializingCacheProvider =
                 stubCacheProvider.AsyncFor<StubSerialized<ResultPrimitive>>().WithSerializer(stubTResultSerializer);
-            ResultPrimitive fromCache = await serializingCacheProvider.GetAsync(key, CancellationToken.None, false);
+            (bool cacheHit, ResultPrimitive fromCache) = await serializingCacheProvider.TryGetAsync(key, CancellationToken.None, false);
 
-            deserializeInvoked.Should().Be(false);
+            cacheHit.Should().BeFalse();
+            deserializeInvoked.Should().BeFalse();
             fromCache.Should().Be(default(ResultPrimitive));
         }