+++ src/Polly.SharedSpecs/Helpers/Caching/StubCacheProvider.cs
@@ -26,20 +26,20 @@ namespace Polly.Specs.Helpers.Caching
 
         private readonly Dictionary<string, CacheItem> cachedValues = new Dictionary<string, CacheItem>();
 
-        public object Get(string key)
+        public (bool, object) TryGet(string key)
         {
             if (cachedValues.ContainsKey(key))
             {
                 if (SystemClock.DateTimeOffsetUtcNow() < cachedValues[key].Expiry)
                 {
-                    return cachedValues[key].Value;
+                    return (true, cachedValues[key].Value);
                 }
                 else
                 {
                     cachedValues.Remove(key);
                 }
             }
-            return null;
+            return (false, null);
         }
 
         public void Put(string key, object value, Ttl ttl)
@@ -50,9 +50,9 @@ namespace Polly.Specs.Helpers.Caching
         #region Naive async-over-sync implementation
 
         // Intentionally naive async-over-sync implementation.  Its purpose is to be the simplest thing to support tests of the CachePolicyAsync and CacheEngineAsync, not to be a usable implementation of IAsyncCacheProvider.  
-        public Task<object> GetAsync(string key, CancellationToken cancellationToken, bool continueOnCapturedContext)
+        public Task<(bool, object)> TryGetAsync(string key, CancellationToken cancellationToken, bool continueOnCapturedContext)
         {
-            return Task.FromResult(Get(key));
+            return Task.FromResult(TryGet(key));
         }
 
         public Task PutAsync(string key, object value, Ttl ttl, CancellationToken cancellationToken, bool continueOnCapturedContext)